% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{tavella_randall}
\alias{tavella_randall}
\title{Generate a sequence based on the Tavella and Randall method}
\usage{
tavella_randall(x_0, alpha, x_min, x_max, n_grid)
}
\arguments{
\item{x_0}{Numeric. The central value around which the sequence is generated.}

\item{alpha}{Numeric. A scaling parameter that controls the shape of the sequence.}

\item{x_min}{Numeric. The minimum value of the sequence.}

\item{x_max}{Numeric. The maximum value of the sequence.}

\item{n_grid}{Integer. The number of intervals in the generated sequence
(\code{n_grid + 1} nodes).}
}
\value{
A numeric vector of length \code{n_grid + 1} containing the generated sequence.
}
\description{
This function generates a sequence of numbers following the Tavella and Randall method,
which applies a combination of hyperbolic sine transformations to create a smooth sequence
between a specified minimum and maximum value.
}
\details{
The function divides the interval \code{[0, 1]} into \code{n_grid}
equally spaced points. At each point, it computes the corresponding value based on the
Tavella and Randall method, which uses the hyperbolic sine function to ensure a smooth
transition between \code{x_min} and \code{x_max}.
}
