% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{american_option_2d}
\alias{american_option_2d}
\title{American Option 2D}
\usage{
american_option_2d(
  s_0,
  x_0,
  k,
  tau,
  r_d,
  r_f,
  q,
  sigma_s,
  sigma_x,
  rho,
  type,
  s_min,
  s_max,
  x_min,
  x_max,
  n_s,
  n_x,
  n_t,
  alpha,
  lambda,
  tolerance
)
}
\arguments{
\item{s_0}{Stock spot price}

\item{x_0}{FX spot price}

\item{k}{Strike price}

\item{tau}{Time to expiry}

\item{r_d}{Risk-free rate (domestic)}

\item{r_f}{Risk-free rate (foreign)}

\item{q}{Dividend yield}

\item{sigma_s}{Volatility of the stock}

\item{sigma_x}{Volatility of the FX}

\item{rho}{Correlatin between the stock and the FX}

\item{type}{Either "call" or "put"}

\item{s_min}{Min value of the stock prices grid}

\item{s_max}{Man value of the stock prices grid}

\item{x_min}{Min value of the FX prices grid}

\item{x_max}{Max value of the FX prices grid}

\item{n_s}{Size of stock grid for finite difference grid}

\item{n_x}{Size of FX grid for finite difference grid}

\item{n_t}{Size of time grid for finite difference grid}

\item{alpha}{Parameter that defines the uniformity of the grid}

\item{lambda}{Penalty parameter greater than 1}

\item{tolerance}{Error tolerance}
}
\description{
This function evaluates an American-style option on a common stock in a foreing currency using finite differences.
}
