% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{american_option_lv}
\alias{american_option_lv}
\title{American Option 1D (local volatility, penalty method)}
\usage{
american_option_lv(
  s_0,
  k,
  tau,
  r_d,
  q,
  sigma,
  type,
  s_min,
  s_max,
  n_s,
  n_t,
  lambda,
  tolerance
)
}
\arguments{
\item{s_0}{Stock spot price.}

\item{k}{Strike price.}

\item{tau}{Time to expiry (in years).}

\item{r_d}{Risk-free rate (domestic).}

\item{q}{Dividend yield.}

\item{sigma}{Local volatility matrix sampled on the grid.}

\item{type}{Either "call" or "put".}

\item{s_min, s_max}{Min/Max of the underlying prices grid.}

\item{n_s}{Number of intervals in the asset grid (n_s + 1 nodes).}

\item{n_t}{Number of time steps.}

\item{lambda}{Penalty parameter (> 1).}

\item{tolerance}{Relative error tolerance for the penalty iterations.}
}
\description{
Prices an American option on a single equity with local volatility using operator splitting
and a penalty-projection method for the early exercise constraint.
}
