% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{european_option_lv}
\alias{european_option_lv}
\title{European Option Local Volatility 1D}
\usage{
european_option_lv(s_0, k, tau, r_d, q, sigma, type, s_min, s_max, n_s, n_t)
}
\arguments{
\item{s_0}{Stock spot price}

\item{k}{Strike price}

\item{tau}{Time to expiry}

\item{r_d}{Risk-free rate (domestic)}

\item{q}{Dividend yield}

\item{sigma}{Local volatility matrix}

\item{type}{Either "call" or "put"}

\item{s_min}{max value of the underlying prices grid}

\item{s_max}{min value of the underlying prices grid}

\item{n_s}{Size of asset grid for finite difference grid}

\item{n_t}{Size of time grid for finite difference grid}
}
\description{
This function evaluates an European-style option on a common stock with local volatility using finite differences.
}
